# Conflicts:
#   src/peersafe/app/table/impl/TableSync.cpp
#   src/ripple/app/ledger/impl/LedgerMaster.cpp
#   src/ripple/app/misc/NetworkOPs.cpp

+++ src/test/vm/vm_test.cpp
@@ -19,12 +19,15 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
 #include <string>
 #include <vector>
+#include <thread>
+#include <fstream>
 
 #include <ripple/beast/unit_test.h>
 #include <ripple/basics/StringUtilities.h>
 #include <boost/algorithm/string.hpp>
 
-#include "FakeExtVM.h"
+#include <test/vm/FakeExtVM.h>
+//#include "FakeExtVM.h"
 
 /*
 * usage:
@@ -189,7 +192,7 @@ private:
 		size_t t1_end = (runCodes.size() / 2) - 1;
 		std::thread t1([this,&worker, &runCodes, t1_start, t1_end]() {
 			
-			long idx = 1;
+			uint8_t idx = 1;
 			for(size_t i = t1_start; i < t1_end; i++) {
 				bytes code;
 				code.assign(runCodes[i].begin(), runCodes[i].end());
@@ -205,7 +208,7 @@ private:
 		size_t t2_end = runCodes.size();
 		std::thread t2([this, &worker, &runCodes, t2_start, t2_end]() {
 			std::this_thread::sleep_for(std::chrono::milliseconds(20));
-			long idx = 1;
+			uint8_t idx = 1;
 			for(size_t i = t2_start; i < t2_end; i++) {
 				bytes code;
 				code.assign(runCodes[i].begin(), runCodes[i].end());
